Extract LSL subclasses so it isn't a Python module
