+++ mentalabMobileApi/src/main/java/com/mentalab/service/lsl/ChannelFormat.java
@@ -1,0 +1,8 @@
+package com.mentalab.service.lsl;
+
+public class ChannelFormat {
+
+    public static final int float32 = 1;
+    public static final int int16 = 5;
+    public static final int int32 = 4;
+}
